requirements.txt
absl-py==1.2.0
anyio==3.6.2
api==0.0.7
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
astunparse==1.6.3
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.12.0
bleach==5.0.1
cachetools==5.2.0
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cycler==0.11.0
Cython==0.29.24
debugpy==1.6.5
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
exceptiongroup==1.1.0
fastapi==0.86.0
fastjsonschema==2.16.2
filelock==3.9.0
flake8==5.0.4
Flask==2.2.2
Flask-Cors==3.0.10
flatbuffers==22.11.23
fonttools==4.38.0
gast==0.2.2
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grade==3.0.1
grpcio==1.49.1
h11==0.14.0
h5py==2.10.0
idna==3.4
imageio==2.19.0
imgaug==0.4.0
importlib-metadata==4.2.0
importlib-resources==5.10.2
iniconfig==2.0.0
ipykernel==6.16.2
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.23.4
jupyter_client==7.4.8
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.4
mypy==0.991
mypy-extensions==0.4.3
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.6
networkx==2.6.3
nose==1.3.7
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.21.6
oauthlib==3.2.2
opencv-python==4.7.0.68
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.10.3
pickleshare==0.7.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
platformdirs==2.6.2
pluggy==1.0.0
prometheus-client==0.15.0
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.2
pyflakes==2.5.0
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==7.2.1
pytest-flake8==1.1.1
pytest-mypy==0.10.3
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.6
PyWavelets==1.3.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
Send2Trash==1.8.0
Shapely==1.8.5.post1
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
starlette==0.20.4
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorflow-estimator
tensorflow-gpu
tensorflow-io-gcs-filesystem
tensorflow-serving-api
tensorflow-serving-api-gpu
termcolor==2.0.1
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2021.11.2
tinycss2==1.2.1
tomli==2.0.1
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
widgetsnbextension==4.0.5
wincertstore==0.2
wrapt==1.14.1
zipp==3.8.1
